requirements.txt
customtkinter
Pillow
pyserial
pyinstaller
